+++ plugins/offline/offline.js
@@ -320,6 +320,10 @@ Offline.prototype.checkin = function checkin(job, done) {
     done(err);
   });
 
+  // Request from the offline queue need to have the noBlock set to true to
+  // ensure that the check-in is forced in the library system.
+  data.noBlock = true;
+
   // Send request to FBS.
   self.bus.emit('fbs.checkin', data);
 };
@@ -364,6 +368,10 @@ Offline.prototype.checkout = function checkout(job, done) {
     self.bus.emit('logger.offline', 'error: ' + err.message);
     done(err);
   });
+
+  // Request from the offline queue need to have the noBlock set to true to
+  // ensure that the check-out is forced in the library system.
+  data.noBlock = true;
 
   // Send request to FBS.
   self.bus.emit('fbs.checkout', data);